Edit studentModel and studentController to allow professors to access only their students

+++ backend/src/models/studentModel.js
@@ -1,22 +1,22 @@
 import pool from '../config/db.js'
 
-export const getAllStudents = async () => {
-    const result = await pool.query('SELECT * FROM student')
+export const getAllStudents = async (userId) => {
+    const result = await pool.query('SELECT * FROM student s JOIN professor_student ps ON s.id = ps.student_id WHERE ps.professor_id = $1', [userId])
     return result.rows
 }
-export const getStudentById = async (id) => {
-    const result = await pool.query('SELECT * FROM student WHERE id = $1', [id])
+export const getStudentById = async (userId, id) => {
+    const result = await pool.query('SELECT * FROM student s JOIN professor_student ps ON s.id = ps.student_id WHERE ps.professor_id = $1 and s.id = $2', [userId, id])
     return result.rows[0]
     }
 export const addStudent = async (nom, prenom, note) => {
-    const result = await pool.query('INSERT INTO student (nom, prenom, note) VALUES ($1, $2, $3) RETURNING *', [nom, prenom, note])
     return result.rows[0]
 }
-export const updateStudent = async (id, nom, prenom, note) => {
-    const result = await pool.query('UPDATE student SET nom = $1, prenom = $2, note = $3 WHERE id = $4 RETURNING *', [nom, prenom, note, id])
+export const updateStudent = async (nom, prenom, note, id, userId) => {
+    const result = await pool.query('UPDATE student s SET nom = $1, prenom = $2, note = $3 FROM professor_student ps WHERE s.id = ps.student_id  AND s.id = $4 AND ps.professor_id = $5 RETURNING s.*;', [nom, prenom, note, id, userId])
+
     return result.rows[0]
 }
-export const deleteStudent = async (id) => {
-    const result = await pool.query('DELETE FROM student WHERE id = $1 RETURNING *', [id])
+export const deleteStudent = async (userId, id) => {
+    const result = await pool.query('DELETE FROM student s USING professor_student ps WHERE s.id = ps.student_id AND s.id = $1 AND ps.professor_id = $2 RETURNING s.*', [id, userId])
     return result.rows[0]
 }
